Fix OutOfMemoryError caused by the Filter Media process when filtering some large TIF files
When thumbnail.blurring = true in dspace.cfg it can cause the blurring process to throw an OOME for large TIF files. Now, if an OOME's thrown, a low quality thumbnail is generated instead of the process failing that resulted in no subsequent items with bitstreams to be processed. Same behaviour when thumbnail.hqscaling = true that causes an OOME.